+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Which Angel Ring wing you have is now a Data Attachment
 	- It was previously a packet that modified a static `Map<UUID, Type>` which is gross
+- The Ender Quarry's beam renders while the EQ itself is off-screen now
 
 ### Fixed